requirements.txt
Flask
requests
python-dotenv
sentence-transformers
scipy
convex
google-generativeai
uvicorn
fastapi
fastapi-cors
requests
google-api-python-client
lxml